fix(perf): 优化 CLS、SEO、文案一致性与首页组件拆分
背景：
- PageSpeed Insights 显示 CLS 偏高，主要由广告位和图片缺少尺寸占位导致。
- 首页更新弹窗 500ms 弹出太快，打断用户主路径。
- 首页广告位紧贴 stats 区域，进入内容流过早。
- index.html 缺少 meta description、OG、Twitter Card、canonical。
- README 隐私表述与实际 AdSense 使用存在矛盾。
- HomePage.vue 已达 1048 行，可维护性下降。

实现：
- AdsenseSlot 添加 min-height: 144px 和 max-height: 280px 稳定占位。
- 首页永雏塔菲展示图补充 width/height 属性。
- 结果页主视觉图补充 width/height 属性和 aspect-ratio: 1。
- 首页广告位从 stats 后移至 testimonials 后。
- 更新弹窗延迟从 500ms 调整为 3000ms。
- index.html 添加 meta description、Open Graph、Twitter Card、canonical。
- README 修正隐私表述，明确说明 Google AdSense 第三方广告服务。
- README 角色数量描述统一为 105 个（含隐藏角色）。
- 首页拆分为 5 个子组件：HeroSection、StatsSection、FeaturedStream、
  TestimonialsSection、UpdatePopup。

验证：
- npm run build 构建通过，无 TS 错误。

影响：
- CLS 预期显著改善，广告和图片不再造成布局偏移。
- 社交分享预览将展示正确的 OG 图和描述。
- README 隐私表述与站点现实一致。
- HomePage.vue 从 1048 行降至约 200 行，可维护性提升。
