+++ README.md
@@ -42,7 +42,7 @@ ACG Type Indicator
 - **105 位角色库**：当前包含 105 位角色，涵盖 60+ 部热门作品，包括 BanG Dream!（Ave Mujica / MyGO）、孤独摇滚！、鸣潮、明日方舟、轻音少女、我推的孩子、Re:从零开始的异世界生活、败犬女主太多了！、辉夜大小姐想让我告白、青春猪头少年、魔法少女小圆、某科学的超电磁炮、名侦探柯南、EVA、东方Project、VOCALOID、原神、崩坏：星穹铁道、紫罗兰永恒花园、四月是你的谎言、葬送的芙丽莲、间谍过家家、刀剑神域、Fate/stay night、电锯人、少女乐队的呐喊等，持续扩充中。
 - **可视化交互**：16personalities 风格的交互式倾向滑块，直观展现你的思维倾向。
 - **一键分享**：精美的结果图报表，支持一键导出 PNG 海报分享给同好。
-- **纯前端架构**：无后端、无注册、无数据收集，测算过程全部在本地浏览器完成，极致极速且保护隐私。
+- **纯前端架构**：无后端、无注册，测试结果全部在本地浏览器完成计算。站点含第三方广告服务（Google AdSense），详见下方隐私说明。
 
 ## 🛠️ 技术栈
 
@@ -127,7 +127,7 @@ src/
 |:-----|:-----|
 | `src/data/questions.json` | 39 道情境式题目 — 维度、原型权重、场景标签 |
 | `src/data/archetypes.json` | 8 个角色原型 — 名称、描述、亮点、短板 |
-| `src/data/characters.json` | 105 个角色条目 — 角色代码、MBTI 映射、标签、六维向量 |
+| `src/data/characters.json` | 105 个角色条目（含隐藏角色） — 角色代码、MBTI 映射、标签、六维向量 |
 | `src/data/characterVisuals.json` | 角色视觉配置 — 立绘、色彩、主题 |
 | `src/data/characterProbabilities.json` | 角色命中概率 — 基于人群统计的先验分布 |
 
@@ -240,6 +240,7 @@ git push origin v0.2.0
 
 - 本工具为**纯前端运行**架构（无后端服务器）。
 - 您的所有测试过程、点击数据与最终生成的结果海报，均只在您的**本地浏览器**中计算生成。我们**不会**收集、上传或存储您的任何答题数据或个人隐私信息。
+- 本站页面包含由 **Google AdSense** 提供的第三方广告服务，广告脚本可能使用 Cookie 进行个性化广告投放。这属于第三方服务行为，不属于本项目的数据收集范畴。如需了解广告隐私政策，请参阅 [Google 隐私权政策](https://policies.google.com/privacy) 与 [广告设置](https://adssettings.google.com/)。
 
 ### 测试结果声明
