+++ index.html
@@ -5,7 +5,25 @@
     <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
     <meta name="theme-color" content="#33a474" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>ACGTI - ACG Type Indicator</title>
+    <title>ACGTI - ACG Type Indicator | 二次元角色原型测试</title>
+    <meta name="description" content="ACGTI 是以 MBTI 为基础的二次元角色原型测试。回答 39 道情境式问题，获得唯一命中的角色代码，解锁你的二次元人格原型。纯前端运行，无需注册，105 位角色等你来测。" />
+    <link rel="canonical" href="https://acgti.tianxingleo.top/" />
+
+    <!-- Open Graph -->
+    <meta property="og:type" content="website" />
+    <meta property="og:url" content="https://acgti.tianxingleo.top/" />
+    <meta property="og:title" content="ACGTI - ACG Type Indicator" />
+    <meta property="og:description" content="以 MBTI 为基础的二次元角色原型测试。回答情境式问题，获得唯一命中的角色代码，解锁你的二次元人格原型。" />
+    <meta property="og:image" content="https://acgti.tianxingleo.top/og-image.png" />
+    <meta property="og:locale" content="zh_CN" />
+    <meta property="og:site_name" content="ACGTI" />
+
+    <!-- Twitter Card -->
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta name="twitter:title" content="ACGTI - ACG Type Indicator" />
+    <meta name="twitter:description" content="以 MBTI 为基础的二次元角色原型测试。回答情境式问题，获得唯一命中的角色代码。" />
+    <meta name="twitter:image" content="https://acgti.tianxingleo.top/og-image.png" />
+
     <script
       async
       src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-1224030144383381"
